+++ screens/Player.js
@@ -1,31 +1,19 @@
-import React from 'react';
-import { StyleSheet, Text, View } from 'react-native';
+import React, {useState} from 'react';
+import { StyleSheet,View } from 'react-native';
 import Ionicons from 'react-native-vector-icons/Ionicons'
-import Menu from '../components/Menu';
-import { Container, Header } from 'native-base';
+import Headers from '../components/Headers'
+
+
 
 export default function Player(props){
-  // static navigationOptions = {
-  //   drawerIcon: () => (
-  //     <View style={styles.icon}>
-  //       <Ionicons
-  //         name="logo-game-controller-b"
-  //         size={24}
-  //         style={styles.iconcolor}
 
-  //       />
-  //     </View>
-  //   )
-  // }
-
-  
     return (
-      <Container style={styles.view} >
-        <Header style={{ marginTop: 26 }}>
-          <Menu navigation={props.navigation} />
-          <Text style={styles.text1}>Player </Text>
-        </Header>
-      </Container>
+     <View style={{flex: 1}}>
+       <Headers  
+       name ='PLayer'
+       navigation={props.navigation}
+ />
+     </View>
     );
   }
 
@@ -43,15 +31,7 @@ export default function Player(props){
     }
 
 const styles = StyleSheet.create({
-  view: {
-    flex: 1,
-
-  },
-  text1: {
-    textAlign: 'center',
-    fontSize: 20,
-    color: "white"
-  },
+  
   icon: {
     fontSize: 15,
     textAlign: 'center',
